NHibernate Saga persister does not work with Oracle
The issue is that the Oracle driver does not support both TransactionScope and NH transaction, see http://stackoverflow.com/q/6870901/90882 for more info.
So we now checking if there is currently an active ambient transaction and if there is we do not create a NH transaction.
We also disable AutoFlush and control Flush manually.